[IMP] web_grid: continue editing when hitting "enter" on grid cell
Before this commit, when the user hits on "Enter" key when he has
finished to edit a grid cell, the edition is finished, that is, it
does not move into another cell, it just stops the edition.

This commit allows the user to continue to edit when he hits on
"Enter" key, that is, when the user hits on "Enter" key, the cell
below the one edited will be in edit mode and once the user will be
in the last row and hits "Enter" then the cell in the first and in
the same column of the one edited will be in edit mode.

task-2618994

Part-of: odoo/enterprise#42680