[IMP] account_reports: improve file generation errors handling
The file export generation blocks with user errors when some fields are not filled or wrongly computed.
We want to avoid such blocking behavior, and allow the user to still generate the report, even if it's wrong.
The errors that occurred are still notified to the user in a new wizard that allows him/her to generate the report anyway.
This wizard also provides a list of all the errors that occurred during the generation process, instead of blocking at the first one.

closes odoo/enterprise#42385

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>