[ADD] *: online table booking
*:appointment,pos_preparation_display,pos_restaurant_appointment,
website_appointment

This module adds the possibility to book a table online from the
website. The booking is then visible in the POS.

When a new table is created in the PoS, a linked resource is created
in the appointment application, which allows the user to add these
resources to appointment_types, authorising visitors to reserve tables
online.

In pos_config, you need to choose an appointment_type via which the
appointments will be retrieved

A websocket has been integrated to retrieve new bookings instantly
from the PoS, which are displayed directly on the tables.
When an event (appointment) is created, only that new event is sent
thru the bus as notification.

Some improvements have also been made to the pos_preparation_display
module on the websocket part.

closes odoo/enterprise#42489

Related: odoo/odoo#124850
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>