[FIX] web_enterprise: fix iOS app icon for PWA
This commit allows the PWA to use an appropriate icon for the
iOS installation. Before this commit, an iPhone would install
the app using the base favicon, with black background and no
padding. Now, a custom icon made specifically for the needs of
iOS is used, with the apple-touch-icon added in the template.

A test has been added to check the correct behavior of the route
of the apple-touch-icon image.

The behavior of Safari regarding PWA icons is not standard, and
the documentation (1) related to the installation of a website only
refer to this apple-touch-icon for iOS devices. Without this custom
icon, a zoomed version of the favicon was used when installing
the app on those devices.

(1): https://developer.apple.com/library/archive/documentation/AppleApplications/Reference/SafariWebContent/ConfiguringWebApplications/ConfiguringWebApplications.html

closes odoo/enterprise#42661

X-original-commit: 5c48d99497845542541c10dc0f6f659eb1d5c416
Signed-off-by: Romain Estievenart (res) <res@odoo.com>