[FIX] *: ensure tour tips are correctly translated
*: account_accountant, documents, documents_account, helpdesk,
hr_expense_extract, hr_payroll, industry_fsm, industry_fsm_sale,
planning, sale_renting, sale_subscription, sign, social,
stock_barcode_mrp_subcontracting, timesheet_grid, web_studio

In odoo/odoo#111103 the tour system was rewritten. The previous tour
system used to depend on the `root.widget` js module, and this module
was an async module that indirectly depended on `session_bind` which
would load the translations, meaning that the js module definition code
of the tours would only run after the translations were loaded. This is
no longer the case with the new tour system, this means that the module
definition code is executed as soon as the dependencies of that module
are fulfilled, which is generally before the translations are loaded,
causing most tour tips to not be translated.

The community commit adds a legacy module that can be imported for the
side-effect of loading the translations, and this commit imports that
module in the onboarding tours, which causes them to be correctly
translated once again.

closes odoo/enterprise#42655

X-original-commit: 02ac0725ba5708bd057b0edf63d61e4da60be2eb
Related: odoo/odoo#125284
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>