[FIX] knowledge: stop using allow_end_on_form
`allow_end_on_form=True` is not safe and not the default.
It should only be opted in when tere's no other choice.

It introduces undeterministic errors as if a tour ends on a form view that is
not saved, the test can be markes as finished and successfull while some threads
are still running.

The proposed solution is to add steps to leave the form view in every Knowledge
tour, and stop using `allow_end_on_form=True`

task-3345004

closes odoo/enterprise#42943

X-original-commit: 82faeacbed2c14d1384c022b6d9094ee9ae7bb40
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>