[FIX] mrp_workorder: remove workaround for obscure ORM bug
The ORM issue was a new record with an x2many field value being real
records instead of new records, which violates one of the new records'
invariants.  The code was computing the intersection between two
recordsets, and because of the bug, one of them was new records and the
other one was real records.

Since the bug is now fixed, the workaround has to go away.

closes odoo/enterprise#42445

Related: odoo/odoo#124612
Signed-off-by: Raphael Collet <rco@odoo.com>