[FIX] knowledge: fix dead link on action helper
The link that allows users to create a new article from the action helper
of the Knowledge editor is currently broken: When the user clicks on it,
nothing happens. This commit will address that issue by attaching to that
link a listener that will allow the creation of an article.

Steps to reproduce the issue:
1. Go on Knowledge
2. Delete all articles
3. Open the Knowledge editor
4. Click on the "Create an article" link of the action helper
=> Nothing happens

TO BE: When the user clicks on the link, the system should create a new
private article and open it right away.

Related: https://github.com/odoo/enterprise/pull/34788

task-3367208

Part-of: odoo/enterprise#42387