[FIX] website_helpdesk: traceback resolve for template 'Ticket Received'
Steps to produce the traceback
==============================
1. Install website_helpdesk.
2. Open Website --> Site menu --> Pages --> Click on ticket received page title.

Technical
=========
If the 'ticket' variable value is null then the same value is being used in
t-attf-href="/helpdesk/ticket/#{ticket.id}" that causes the view to be broken
and causes the traceback.

After this Commit
=================
The conditional rendering of the tag based on 'ticket' value resolves the traceback.

Task-3366890

closes odoo/enterprise#43873

X-original-commit: 99c031ab7ee2db68598543de733a0cd690808d97
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>