[FIX] approvals: active_id is passed as string
When the user tries to access the partners in 'approvals' module from contacts,
where in the context the category_id is 'active_id'.

To reproduce the issue:
- Install 'Approvals'
- In Approvals, go to Configuration > Approvals Type
- Select any type and change the 'Contact' field to 'Required'
- Save the changes
- Go to My Approvals > New Request
- Click on any category
- In 'Contacts' drop down list click on 'Search More...'
Error: A traceback appear: 'ValueError: Expected singleton: res.partner.category
('a', 'c', 't', 'i', 'v', 'e', '_', 'i', 'd')'

The category_id is set to active_id which is passed as a string due to which
the traceback is generated.

sentry-4237907737

closes odoo/enterprise#42630

X-original-commit: 05a9b9ba832d9919e7964ac7b864864dce064d78
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>