[FIX] account_followup: join_invoice followup report
Before this commit, when a pdf was already generated
from an invoice, the pdf was joined to the followup
reminder, even if 'join_invoices' was set to False
on the followup line.

Steps:

- Create a followup level to send an email
  and set 'join_invoices' field to False
- Create an invoice
- Go to Followup Reports
- Select the customer from invoice above,
  and click on Follo Up button
-> On the wizard, the join_invoices
   checkbox is unchecked but attachment is
   still attached, you can see it if you
   check the checkbox, therefore the pdf
   will be joined to the email.

With this comit, we remove the attachment
if the checkbox is not checked on the
wizard when processing the followup.

opw-3328105

closes odoo/enterprise#42498

Signed-off-by: William André (wan) <wan@odoo.com>