[IMP] sale_subscription: don't tokenize partial payment
Before this commit, all token available were saved on the sale order on confirmation.
This caused an issue when the transaction was a partial payment of the invoice.
No invoice would be created and therefore the next invoice date would be equal to the
start date.

The invoicing cron would run and charge the total amount on the token.
This commit prevent to save the token in case of partial payments.

closes odoo/enterprise#43630

Taskid: 3375151
X-original-commit: 2b86813365b4f584a8045d1dbb4d75706b5724b9
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>