[FIX] account_reports: multicurrency revaluation report: buggy warning
When setting a custom exchange rate through the appropriate filter of the report, a warning is displayed reminding the user custom rates are used in the displayed data. This warning contains a button to reset the rates to their original values. Clicking it caused the custom rates filter to disappear ; it's not what we want. This commit restores the proper reset behavior.

closes odoo/enterprise#42616

X-original-commit: fe2c2c2925f274be74b9f4dc8c084f3f04fa8724
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>