[FIX] knowledge: save field value when before doAction
`askChanges` cannot be called in the `updateBehaviors` process, as it would
create an infinite loop since `updateValue` calls `updateBehaviors` and
`askChanges` calls `commitChanges` which calls `updateValue`.

Furthermore the reason why `askChanges` was put in the `updateBehaviors` process
is dubious, as "rendering a Behavior" is not a good time to mark the record
containing the `htmlField` as `dirty`.

The issue that some content can be lost if a `doAction` is executed through a
Behavior component is caused by the fact that `htmlField.commitChanges` is
called when the `blur` event occur on the editable element, but since `doAction`
can be called by clicking on an element inside the `editable`, the `blur` event
will not be triggered, resulting in the record not being set to `dirty`.

There is still an issue with `doAction` that can be called from code that is not
controlled by `Knowledge` (i.e. embedded views), but this will be the object of
another task: 3372343

The fix in this commit fixes the issue of unsaved content for the `/article`
command block.

previous attempt to fix the issue (implementation reverted):
https://github.com/odoo/enterprise/commit/4134fbdc071ac04c2e74466204ab57ab4eb5acb4

restore tour steps removed from no reason here:
https://github.com/odoo/enterprise/commit/5b6df33e9a404a91230374fbfebd362d88f11552

task-3280704

closes odoo/enterprise#42615

X-original-commit: a0ea883fcc2abebcc08421d7b629c3fde857f6fd
Related: odoo/odoo#125204
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>