[FIX] industry_fsm, project: add a default project
Steps to reproduce:
-------------------
- create a stage without selecting a project.

Issue:
------
The new stage is inaccessible.

Cause:
------
When the project is not defined, the following ORM command is used:
`[6, False, []]` for the `project_ids` field.

Solution:
---------
When we create a project stage (project task type),
it is essential that we link it to a project.
Add the possibility to override `_get_default_project_ids`.
Set default project as first fsm project if not project in
the context.

opw-3322992

closes odoo/enterprise#42904

X-original-commit: bc47e5ddbcb7a21bb82d1b2b996dda04a5f9f465
Related: odoo/odoo#125884
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>