[FIX] mrp_mps: mps button on Bom open wrong record
When search mps by bom, currently we always use "ilike" as operator to
search by name. This leads to wrong result when we search by id, bom
with the "id" in its name will also be shown.
To fix, we don't force to use "ilike". Allow it to change by different
cases.

Task-3337966

X-original-commit: a254ed65f22e1ba4be419e7cae60d1386d5d67d9
Part-of: odoo/enterprise#44991