[FIX] mrp_mps: mps button on Bom open wrong record
When search mps by bom, currently we always use "ilike" as operator to
search by name. This leads to wrong result when we search by id, bom
with the "id" in its name will also be shown.
To fix, we don't force to use "ilike". Allow it to change by different
cases.

Task-3337966

Part-of: odoo/enterprise#42547