[FIX] marketing_automation: update mail schedule date
When we try to update a marketing automation workflow that has emails that should be send after some actions (eg: click, reply, etc), we have to update the schedule date but currently we are trying to set date this way:

process_dt = trace.parent_id.mailing_trace_ids.state_update

But state_update is never defined in the code which caused this issue. To fix the issue the new schedule date is defined according to the correct action.

How to reproduce:
1. Marketing Automation -> Create -> Add new activity
2. Create two emails in the workflow in a such way, one email is parent of another and the child email is triggered when the email is Opened
3. Click on Start -> Launch test -> start the test workflow
4. Edit the interval_number of the child email

closes odoo/enterprise#43546

Opw: 3355807
X-original-commit: aec2cc7e69a1db9ed8d2bd8c66c27b87cb713cc3
Signed-off-by: Matheus Leal Viana (malv) <malv@odoo.com>