[FIX] timesheet_grid: clock w/ timer header TS w/o approval rights
Current behaviour:
If an user doesn't have any employee rights, they can't clock
timesheets using the header timer bar, they are met with an
Access Error, that only approvers can validate a timesheet.

Expected behaviour:
An user with no employee rights should be able to clock timesheets
using the header bar in the Timesheets app.

Steps to reproduce:
- Install Timesheets
- Remove employee rights of Marc Demo
- Login as Marc Demo
- Go to Timesheets > (any of the view you have the timer header,
  so grid/list/kanban)
- Start the timer, add a project
- Add a task (AccessError!)

Reason for the problem:
The AccessError is triggered by the check `_check_can_write`, which
checks for *presence* of the field `'validated'`. All theses views
I've mentioned define this field as invisible, since it is used for
a few display conditions of the items displayed in the views. When
we start the timer, a new front-end `timesheet` record is created,
which uses the fields defined from the view, so `validated` included.
When we add the task, the hook is saving the current state of the
timesheet by initiating a write with the content of the record,
which includes the field `validated` with the value `false`. This
trips up the check, because our current user doesn't have the group
`hr_timesheet.group_hr_timesheet_approver` and the key `validated`
is present.

Fix:
To avoid having OWL send the field for writing, a solution is to set
such field as readonly, since readonly fields are dropped before
sending the write call to the server. The field is invisible all
cases, so it is safe (it can't be modified by the user anyway). The
only specificity is for the kanban-view. *All* fields that are
defined in kanban view are implicitly `force_save='1'` if not
specified otherwise. Since the presence of `force_save` on the field
`validated` would force OWL to send the field for writing, we
override that attribute to `force_save='0'` for the instance where
the `validated` field is in the kanban view.

Affected versions:
- saas-16.2
- saas-16.3
- master

opw-3334853

closes odoo/enterprise#42690

X-original-commit: 2c07696190882584c13d5a8e3d7203a137369578
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>