[IMP] appointment: Gantt view for resource bookings
This adds a customized Gantt view for appointment events
as well as a form view and some related field which
should allow users to more easily manage their bookings
in terms of who/what and when visually. Instead of using
the calendar view, which isn't as well suited for that.

The gantt customizations involve allowing custom default groupings
and group_expand filters specific to the gantt without impacting
other views of the same actions. This allows users to only see
appointments and resources that are relevant to the use of the
gantt view. Meaning we can exclude resources linked to
appointments with capacity management, as these are likely
to involve multiple resources for a single event, which
gantt cannot support properly.

We also pick colors based on attendance. Users can mark
events as attended, independently from the status of the
event. This allows them to be able to tell which attendees
are running late, and potentially give them a call.

task-3101589

Part-of: odoo/enterprise#37312