[FIX] l10n_be_reports: Handle credit note in 281.50 computation
Before this commit, form 281.50 doesn't handle credit notes
in its computation. It was caused because the join
on account_partial_reconcile didn't take into account
credit notes.

e.g.:
Create a vendor bill(1000€) as fees and post it
Pay the vendor bill (1000€)
Create a credit note for this vendor bill (1000€)
Receive the payement from this credit note (1000€)
281.50 will return 1000€ for fees accounts and
1000€ for paid amount.

It's not what we want and this commit fixes this.
e.g.:
Create a vendor bill(1000€) as fees and post it
Pay the vendor bill (1000€)
Create a credit note for this vendor bill (1000€)
Receive the payement from this credit note (1000€)
281.50 will return 0€ for fees accounts and
0€ for paid amount.

closes odoo/enterprise#42747

Task: none
X-original-commit: 23b57a24c9315c16c819120beb2a6d9ce3e3df08
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>