[FIX] knowledge: close sidebar when opening article on mobile devices
Currently, the sidebar no longer closes on mobile devices when the user
selects or creates an article. The user needs to manually close the menu
to view the article which is not really convenient.

Steps to reproduce the issue:
1. Go to Knowledge using a mobile phone
2. Toggle the sidebar
3. Click on an article or create a new article ("+" sign)
=> The sidebar is kept open.

TO BE: The sidebar should close itself when the user selects or creates
an article. That way, people will not have to do additional clicks to
read or edit the article they want.

Reference: https://github.com/odoo/enterprise/pull/34788

task-3368916

closes odoo/enterprise#42454

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>