[FIX] knowledge: fix search favorites tour
Add a step to ensure that the first kanban view is fully loaded before trying to
insert the next one in `knowledge_items_search_favorites_tour`, because it could
cause an undeterministic error as the view was not inserted when trying to
insert the second one too fast.

Change the triggers to ensure that a view is loaded to be more explicit about
what should be loaded:
`o_knowledge_embedded_view` is only present when the embedded view is loaded by
the manager (and it could happen after the embedded_view Behavior was loaded
(see lazy loading)).

Technical issue explanation:
- Rendering of Behaviors is asynchronous in regards to the execution of the
`/command`.
- Since [1], Behaviors are rendered in parallel.
- This means that the `preserveCursor` call for 2 `/command` that are executed
  in succession faster than the time it takes for Owl to mount the result of
  those commands can be done on the same node, let say `p`.
- The `insert` command of the editor, in case the insertion has to be done in
  between content, will split the editor node with `splitElement`.
- `splitElement` creates 2 clones of the original targetted node, and removes the
  original.
- This means that one of the `preserveCursor` has registered a position inside
  a node that will be removed by the `insert` caused in the process involving
  the other `preserveCursor` and therefore will fail, resulting in only one of
  the two `insert` succeeding.

This issue is not humanly realistic, because the user would have to type two
commands in the same node faster than it takes for Owl to mount the first
Behavior, and currently Behaviors are rendered without heavy `onWillStart` rpc
calls, because we want the user to have a visual feedback as fast as possible.

[1]: https://github.com/odoo/enterprise/commit/b25f4c9ea93f0b7d357fea71d84132efbb6bb234

task-3373322

closes odoo/enterprise#42958

X-original-commit: a00b24f473b2df4066925c2d3239aaada91b0321
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>