[FIX] sale_subscription: Upselling invoicing
Step to reproduce:
- Create a sub with a pre-invoiced recurring and a post-invoice recurring product
- Invoice the sub
- Upsell the Sub with 1 qty for both product
- Invoice the upsell
- Add delivered qty to the post-invoice product
- Invoice the upsell

Current behaviour:
- The second invoice for the upsell also invoice the line that
have already been invoiced during the first invoice.
This is because subscription module add recurring line even if
they are not in the case of an upsell.

Behviour after the change:
- The second invoice only invoice the post-invoice line.
The sale `_get_invoiceable_line` is now also used for upsell
as they don't need recurring behaviour.

opw-3279273

closes odoo/enterprise#43421

X-original-commit: 8063331b0815783178c59200c378efd548726365
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Florian Damhaut (flda) <flda@odoo.com>