[IMP] web_enterprise: adapt tests to changes in community
The PR https://github.com/odoo/odoo/pull/124068/files changes
the action and menu services so that each menu item changes will
(almost) immediately displays an empty screen, to provide some quick
feedback to the user, before showing the actual action, once it is
loaded.

As a result, a few tests have to be adapter, since the actual action is
displayed a tick later in tests.

closes odoo/enterprise#43038

Related: odoo/odoo#124068
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>