[FIX] account_reports : General Ledger: only automatically unfold all when no line was manually unfolded
Steps to reproduce:

Go to general ledger report in the accounting module.

Unfold just one of the lines and print the report.

Current Behaviour:
The pdf printed will have all the lines unfolded.

Expected Behaviour:
The pdf should have unfolded only those lines which were unfolded in the report screen. We only want the report to automatically unfold all of its lines in the export when no line was manually unfolded in the UI.

OPW-3355960

closes odoo/enterprise#42624

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>