[FIX] account_accountant: lock date redirects to nothing
To reproduce
- Ensure there are no draft entries
- Ensure there are only bank transactions that are marked as "to check" but not reconciled
- Try to set a lock date for all users after the date of the "to check" transaction
- A redirect warning is displayed, click "Show Unreconiled Bank Statement Lines"
- You are redirected to the bank recon, but no transactions are displayed

The cause:
The "Not Matched" filter has a domain ('to_check', '=', false). So even though there is a domain on the unreconciled transaction id's, the "Not Matched"  filter excludes them.

The Solution:
Remove the default "Not Matched" filter

The related community PR addresses the issue of HTML displayed in the no Content Helper (when no records are visible)

Task-3366470

closes odoo/enterprise#42641

Related: odoo/odoo#125254
Signed-off-by: Laurent Smet <las@odoo.com>