[IMP] planning : show open shifts when filtering
Before this commit, when the user was performing a search on department
for example, the open shifts were filtered accordingly. This could be
unconvienient when trying to dispatch open shift among the teams, in the
example of the departement search, since they are not shown.

After this commit, The open shift are shown during a search on at least
one of those field : resource, department, manager, job title.
Also, when filtering on 'human' or 'material', all the slots linked to
no role or to a role having at least one resource that is
'human'/'material' (depending on the filter) is shown.

closes odoo/enterprise#42700

Taskid: 3358602
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>