[IMP] knowledge: reorder html_field file
Reorder the HtmlField in preparation for the refactoring. Existing code is
pre-moved inside intermediate functions in order to split `updateBehaviors`
which is becoming quite long and complicated. This will allow the actual
refactoring changes to be more distinguishable.

There are now 3 parts of the method:
1- `_prepareBehaviorAnchor` that will configure the `anchor` in which the Owl
  App should be mounted.
2- `_prepareBehaviorProps` that will parse the props of the Behavior and create
  the object passed to Owl as props for the main component of the App.
3- `_updateBehaviors` is still the main control loop to mount `Behaviors`.

There is no logic change in this commit, it is only moving code. See the REF
commit for the actual improvements.

Move the `updateBehaviors` method as if they were named `mountBehaviors` in
advance, for a better diff management in the REF commit.

task-3383561

Part-of: odoo/enterprise#43087