[FIX] account_saft: Handle partner contacts in a sensible way.
At the moment, the way in which the SAF-T template populates partner
contacts in the XML's Contact->ContactPerson nodes makes no sense at
all. The current behaviour is that for each partner, we create a
ContactPerson node for itself and for each of its children.

That means that if you have n contacts at a partner company, you will
end up with n+1 ContactPerson nodes (one for the company, and one for
each of its employees).

Similarly, if your own company has n employees, you will have n+1
ContactPerson nodes in the SAF-T header, one for your company and one
for each of your employees.

This behaviour is absurd for several reasons.
- The government has no use for more than one contact person at your
  company and at your partner companies. So why list them all?
- If the partner is a company, it makes no sense to put it as its own
  ContactPerson. ContactPersons must be natural persons.
- The ContactPerson node is optional in the SAF-T spec (except when you
  need to provide a ContactPerson for your own company in the SAF-T
  header). So, when we don't have a contact person for a particular
  partner, we should just leave that node blank.

This commit tries to handles this all in a more sensible way:
- If the partner is your own company, put all child contacts of the
  company as ContactPersons.
- If the partner is an individual, put him/her as their own
  ContactPerson.
- If the partner is a company, find the active child partner that has
  the lowest ID (generally a good guess for the main PoC at that
  partner) and put them as the ContactPerson for the company.

closes odoo/enterprise#46337

Taskid: 3395016
X-original-commit: f40ff7739d985ad130aea9e696fa14e36af495a5
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>