[MOV] knowledge: wysiwyg templates for behaviors in one file
The wysiwyg template for the FileBehavior was stored along with its OWL
template, which was not consistent with how it is done for other Behaviors.
Fixes this issue by moving it in `knowledge_editor.xml`, which contains all
`blueprint` templates that represent what is needed by the `Behavior` engine to
mount the `Behavior` components (and it will in turn use the OWL template for
that).

task-3383561

Part-of: odoo/enterprise#43087