[IMP] appointment: improve appointment actions
Previously finding appointments was very cumbersome
as one needed to go to the calendar -> appointments
-> figure out they need to click the right custom stat button
   or go to the appointment type settings and click 'Meetings'

To simplify this, we update the kanban to strictly open the relevant
management view based on the appointment type.

Single user: calendar
Multiple users: gantt
Appointment resources with capacity management: list
Appointment resources without capacity management: gantt

This is done via a single action that will return the relevant
action and modify it based on the different criteria.
Accessing the appointment itself from the gantt is now done though
the gantt actions menu, which is modified to look like a regular
button dropdown.

We also add context keys to pass to the custom gantt for
gantt-specific grouping and filtering, as well as useful
defaults to get the relevant default partner instead
of yourself, and to get a default scale that shows all
the appointments by default.

task-3101589

Part-of: odoo/enterprise#37312