[IMP] crm_enterprise, documents: simplifies the systray
The change is related to odoo/odoo#138135 where we simplify the systray and
define the list view as the default view openned from the systray.
Here we override the default for document to open the activity view instead.

We also update the test as we have now only one clickable area for the activity
group instead of multiple one.

Task-3300854

Part-of: odoo/enterprise#48685