[FIX] account_reports: tax unit filter and warning
When using a tax unit on a tax report, the companies used in the tax unit need to be selected on the company selector for it to work.

The problem was that since the refactoring of account reports on owl, the selection of the tax unit did not trigger the selection of companies. (It was done with a manual reload before).

We fix the issue by setting the companies manually when using the tax unit filter.

Also, tax unit warning was not displayed, it was calling the wrong option key tax_unit instead of tax_units (with the S).

X-original-commit: 6a5b1630cd64ca2f64e2744286afa1e86489679b
Part-of: odoo/enterprise#42798