[IMP] *: add dependencies to `display_name` field
The previous commit (community) introduced an optimization to reduce the
number of queries and fields fetched when we call `name_search`. But it
works much better when the dependencies of `display_name` contain field
names used in the calculation (on the same record/model).

Then, to improve the performance and the cache coherency, add `depends`
and `depends_context` depending on the custom `_compute_display_name`.
Add only the first level of dependencies (never traverse relational
field) because only these have a positive impact on the previous
optimization and the cost is very low (see `modified`).
About `depends_context`, we don't include `lang` because (when `_` is
used by example) it is unlikely to get the same display_name in the same
request with two different lang.

closes odoo/enterprise#42599

Related: odoo/documentation#4639
Related: odoo/odoo#122085
Related: odoo/upgrade#4780
Signed-off-by: Raphael Collet <rco@odoo.com>