[FIX] l10n_in_hr_payroll: print a report without set a layout
If a user prints a 'Yearly Salary by Head' or 'Yearly Salary by Employee'
report without selecting a Layout. Traceback will appear

To reproduce the issue:

- Make debugger mode ON.
- Go to Settings > General Settings >Companies and make sure that layout is not
 selected.
- Install the  'l10n_in_hr_payroll' Module.
- Go to the Employees > Reporting > 'Yearly Salary by Head' or Yearly Salary by
 Employee' report.
- Click on 'Add a line' and add an employee, Then after clicking on the 'PRINT'
button  'Configure your document layout ' is open.
- Now click on the 'SAVE' button.

Error:  KeyError: 'start_date'

When printing a 'Yearly Salary by Head' or 'Yearly Salary by Employee' report
for the first time, we have to select a  'Layout' for the report and
after a save 'Layout' it will pass the incorrect 'active_model'. As a result, we
must pass a current 'active_model' through the 'with_context'.

sentry-4247628606

closes odoo/enterprise#43187

X-original-commit: b54cd9b6cfc5c03c0662a2a99d4e8656f8cb4203
Signed-off-by: Kevin Baptiste <kba@odoo.com>