[FIX] website_sale_stock_renting: daterange filter for rental as public
Current behaviour:
As a public user, when filtering in the shop for a rental product
using the datepicker, we are met with a 403 error, because we don't
have access to the Warehouse model.

Expected behaviour:
You shouldn't have access errors when filtering rental products
based on the date range.

Steps to reproduce:
- Install `website_sale_stock_renting`
- Create a rental product, set it's quant on hand to 1, don't allow
  it to be sold when out of stock
- Go to the shop page, edit the website, customize > daterange picker
- Go to an incognito windows (don't log in, so you are a public user)
- Go to "/shop" and filter based on a daterange you get a 403.

Reason for the problem:
The problem is 2 fold:
- lack of read access on the warehouse model, because in
  `_filter_on_available_rental_products` we are getting the
  `qty_available`, which triggers
  `_compute_quantities/_compute_quantities_dict/_get_domain_locations
  ` which does a search on the Warehouse model. The public user
  doesn't have read access to this model. (The module doesn't define
  any access rules).
- lack of read access on the `sale.order.line` model, used in
  `_filter_on_available_rental_products` to check the availability
  of a variant.

Fix:
The module doesn't define any security rules, but instead of creating
brand new access rights, that would grand read access to both `sale.
order.line` and `warehouse` models to all public users, use a `sudo`
at the problematic call site in the controller that is calling
`_filter_on_available_rental_products`.

Affected versions:
- 16.0
- saas-16.1
- saas-16.2
- saas-16.3
- master

opw-3318852

closes odoo/enterprise#42791

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>
Co-authored-by: Feyensv <vfe@odoo.com>