[FIX] website_sale_subscription: fix display of summarize cart
Before this commit, the 'per unit' condition was inserted between the
if tax_ecluded and the else.

So if you had your website tax excluded, you see 2 prices: the tax
excluded and tax included if it is not a subscription.

In case of tax included and subscription, you don't see any prices, but
just the 'per unit'.

Before this commit, after inheritance:

if tax_excluded:
   -> show price tax excl.
if subscription:
   -> show unit
else:
   -> show price tax included

After this commit, after inheritance:

if tax_excluded:
   -> show price tax excl.
else:
   -> show price tax included
if subscription:
   -> show unit

closes odoo/enterprise#42845

X-original-commit: 300f21a3e5a8923a236943dfe0d7b63790598d92
Signed-off-by: Jérémy Kersten <jke@odoo.com>