[FIX] l10n_cl_edi: fixing chili credit note wizard after refactoring of generic credit note wizard
Previous to this commit:
After removal of the "refund_ method" field (which had 3 possible values), the cl credit note wizard is not working as intented anymore.

Each option triggered specific values for "only text correction" field and "SII Reference Code".

Now the values have to be triggered differently.

After this commit:
Here is how it will work now:
Case 1: 

If the the boolean "only text correction" is set as true by the user: 

- Display the  Fields  "Original Text" and "New corrected text"  and set the field "SII reference code" to "2. Corrects Referenced Document Text". 

Case 2:

If the user keeps the boolean false, then the SII reference code should be set by default the "1. Cancels Referenced Document".

closes odoo/enterprise#43327

Task: 3374344
X-original-commit: b459e81368cbbcbbfef4a4cbec8cc0a445c337f0
Signed-off-by: Josse Colpaert <jco@odoo.com>