[FIX] knowledge: fix dropdown menu overlap in embed views
This commit fixes an overlap that could occur when inserting 2 views next to
each other in a knowledge.article.

If you opened the dropdown menu to select more columns (from a list view), it
would get below the columns header of the list view of the next embed view.

This fix assigns automatic z-index management (only for embed views) to the
"thead" element that removes this overlap while keeping the same behavior.

Task-3340979

closes odoo/enterprise#43076

X-original-commit: a71430c0f1cce75c2657ecf729ea27e8c02a05d2
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>