[FIX] account_consolidation: report: fix 'hierarchy' option
The 'hierarchy' filter of the consolidation report was only shown when an account.group was defined. This was due to the fact this option key was initialized by the logic inherited from the parent class in account_reports module.

This was wrong, as 'hierarchy' filter is overridden in consolidation to serve another purpose, and does then not depend on account.group objects. Even though it was not shown in the UI, it was actually enabled by default all the time ; so the choice made here is now to display the filter all the time as well.

The feature was tested, but the tests were poorly written, and did not call the _get_options function of the report, instead forging an options dictionary directly. This was shadowing the bug, since init_filter_hierarchy wasn't called and hence never could reset the option. This commit fixes the tests as well.

closes odoo/enterprise#42870

X-original-commit: dae8b6d66efcf520c0bd51d94160c359f89b2d97
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>