[REF] appointment,knowledge,web_studio: convert wysiwyg.js to owl
Because the goal of the community commit is to remove all dependency
to legacy in the wysiwyg because the wysiwyg has to be loaded on a lot
of form view (through the html_field), the wysiwyg.js is now an Owl
component.

This commit adapt the enterprise code to work with the new wysiwyg.

task-3175256

Part-of: odoo/enterprise#42141